+++ game-frontend/projects/game-frontend/src/app/app.module.ts
@@ -74,7 +74,6 @@ import { Subscription } from 'rxjs';
 import { GameSidebarComponent } from './components/game-sidebar/game-sidebar.component';
 import { TimeSpecialsComponent } from './components/time-specials/time-specials.component';
 import { TimeSpecialService } from './services/time-specials.service';
-import { HideDuplicatedNamePipe } from './pipes/hide-duplicated-name.pipe';
 
 export const APP_ROUTES: Routes = [
   { path: 'login', component: LoginComponent },
@@ -144,7 +143,6 @@ export const APP_ROUTES: Routes = [
     MissionModalComponent,
     GameSidebarComponent,
     TimeSpecialsComponent,
-    HideDuplicatedNamePipe
   ],
   imports: [
     BrowserModule,