+++ game-frontend/projects/game-frontend/src/app/shared/util/date-time.util.ts
@@ -8,8 +8,8 @@ export interface MilisToDaysHoursMinutesSeconds {
 }
 
 export class DateTimeUtil {
-
     private static readonly _LOG: LoggerHelper = new LoggerHelper(DateTimeUtil.name);
+
     /**
      * Converts the input seconds to an object containing days, hours, minutes and seconds
      *
@@ -20,7 +20,7 @@ export class DateTimeUtil {
      * @memberof DateTimeUtil
      */
     public static milisToDaysHoursMinutesSeconds(inputMilis: number): MilisToDaysHoursMinutesSeconds {
-        this._LOG.todo(['In the future move this functionality to DateUtil, and deprecate this class']);
+        this._LOG.warnDeprecated(this.name, '0.8.1', 'ng://OwgeCore/utils/DateUtil::milisToDaysHoursMinutesSeconds(inputMilis)');
         const unixTime = new Date(inputMilis);
         return {
             days: Math.floor((unixTime.getTime() / 1000) / 86400),