+++ game-frontend/projects/owge-core/src/lib/directives/owge-content.directive.ts
@@ -1,0 +1,10 @@
+import { Directive, Input, TemplateRef } from '@angular/core';
+
+@Directive({
+    selector: '[owgeCoreContent]',
+})
+export class OwgeContentDirective<C = any> {
+    @Input() public select;
+
+    public constructor(public templateRef: TemplateRef<C>) { }
+}