+++ game-frontend/projects/owge-core/src/lib/types/date-representation.type.ts
@@ -1,0 +1,15 @@
+
+/**
+ * A date representation as days, hours, minutes and seconds
+ *
+ * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
+ * @since 0.8.1
+ * @export
+ * @interface DateRepresentation
+ */
+export interface DateRepresentation {
+    days: number;
+    hours: number;
+    minutes: number;
+    seconds: number;
+}